requirements.txt
transformers
torch
pdfminer.six
flask
pytesseract
Pillow
docx2txt
scikit-learn
sqlalchemy
sentencepiece
sentence-transformers
